+++ Source/Scene/BlendingState.js
@@ -1,8 +1,10 @@
 /*global define*/
 define([
+        '../Core/freezeObject',
         './BlendEquation',
         './BlendFunction'
     ], function(
+        freezeObject,
         BlendEquation,
         BlendFunction) {
     "use strict";
@@ -24,9 +26,9 @@ define([
          * @type {Object}
          * @readonly
          */
-        DISABLED : {
+        DISABLED : freezeObject({
             enabled : false
-        },
+        }),
 
         /**
          * Blending is enabled using alpha blending, <code>source(source.alpha) + destination(1 - source.alpha)</code>.
@@ -34,7 +36,7 @@ define([
          * @type {Object}
          * @readonly
          */
-        ALPHA_BLEND : {
+        ALPHA_BLEND : freezeObject({
             enabled : true,
             equationRgb : BlendEquation.ADD,
             equationAlpha : BlendEquation.ADD,
@@ -42,7 +44,7 @@ define([
             functionSourceAlpha : BlendFunction.SOURCE_ALPHA,
             functionDestinationRgb : BlendFunction.ONE_MINUS_SOURCE_ALPHA,
             functionDestinationAlpha : BlendFunction.ONE_MINUS_SOURCE_ALPHA
-        },
+        }),
 
         /**
          * Blending is enabled using alpha blending with premultiplied alpha, <code>source + destination(1 - source.alpha)</code>.
@@ -50,7 +52,7 @@ define([
          * @type {Object}
          * @readonly
          */
-        PRE_MULTIPLIED_ALPHA_BLEND : {
+        PRE_MULTIPLIED_ALPHA_BLEND : freezeObject({
             enabled : true,
             equationRgb : BlendEquation.ADD,
             equationAlpha : BlendEquation.ADD,
@@ -58,7 +60,7 @@ define([
             functionSourceAlpha : BlendFunction.ONE,
             functionDestinationRgb : BlendFunction.ONE_MINUS_SOURCE_ALPHA,
             functionDestinationAlpha : BlendFunction.ONE_MINUS_SOURCE_ALPHA
-        },
+        }),
 
         /**
          * Blending is enabled using additive blending, <code>source(source.alpha) + destination</code>.
@@ -66,7 +68,7 @@ define([
          * @type {Object}
          * @readonly
          */
-        ADDITIVE_BLEND : {
+        ADDITIVE_BLEND : freezeObject({
             enabled : true,
             equationRgb : BlendEquation.ADD,
             equationAlpha : BlendEquation.ADD,
@@ -74,7 +76,7 @@ define([
             functionSourceAlpha : BlendFunction.SOURCE_ALPHA,
             functionDestinationRgb : BlendFunction.ONE,
             functionDestinationAlpha : BlendFunction.ONE
-        }
+        })
     };
 
     return BlendingState;